[Refactoring]: Naming Convention & Add final for immutable class
